Activities are now grouped by date on the activities page!
