+++ client-app/src/features/activities/dashboard/ActivityList.tsx
@@ -1,54 +1,31 @@
 import { observer } from 'mobx-react-lite';
-import React, { SyntheticEvent, useState } from 'react';
-import { Link } from 'react-router-dom';
-import { Button, Item, Label, Segment } from 'semantic-ui-react';
+import React, { Fragment } from 'react';
+import { Header, Item, Segment } from 'semantic-ui-react';
 import { useStore } from '../../../app/stores/store';
+import ActivityListItem from './ActivityListItem';
 
 // Styling out the Activity List in the View.Items SemanticUI format.
 export default observer(function ActivityList() {
   // Use the newly created activityStore to replace the previous handler functionality.
   const { activityStore } = useStore();
-  const { deleteActivity, activitiesByDate, loading } = activityStore;
-
-  // Grab the id for the activty that is to be deleted.
-  const [target, setTarget] = useState('');
-  // Delete click event handler.
-  function deleteActivityHandler(e: SyntheticEvent<HTMLButtonElement>, id: string) {
-    // Get the target's name.
-    setTarget(e.currentTarget.name);
-    // Delete it.
-    deleteActivity(id);
-  }
+  const { groupedActivities } = activityStore;
 
   return (
-    <Segment>
-      <Item.Group divided>
-        {activitiesByDate.map((activity) => (
-          <Item key={activity.id}>
-            <Item.Content>
-              <Item.Header as="a">{activity.title}</Item.Header>
-              <Item.Meta>{activity.date}</Item.Meta>
-              <Item.Description>
-                <div>{activity.description}</div>
-                <div>
-                  {activity.city}, {activity.venue}
-                </div>
-              </Item.Description>
-              <Item.Extra>
-                <Button as={Link} to={`/activities/${activity.id}`} floated="right" content="View" color="blue"></Button>
-                <Button
-                  name={activity.id}
-                  loading={loading && target === activity.id}
-                  onClick={(e) => deleteActivityHandler(e, activity.id)}
-                  floated="right"
-                  content="Delete"
-                  color="red"></Button>
-                <Label basic content={activity.category} />
-              </Item.Extra>
-            </Item.Content>
-          </Item>
-        ))}
-      </Item.Group>
-    </Segment>
+    <Fragment>
+      {groupedActivities.map(([group, activities]) => (
+        <Fragment key={group}>
+          <Header sub color="teal">
+            {group}
+          </Header>
+          <Segment>
+            <Item.Group divided>
+              {activities.map((activity) => (
+                <ActivityListItem key={activity.id} activity={activity} />
+              ))}
+            </Item.Group>
+          </Segment>
+        </Fragment>
+      ))}
+    </Fragment>
   );
 });
